# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChapterVerdict": {
        "properties": {
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/DimensionOut"
            },
            "title": "Dimensions",
            "type": "array"
          }
        },
        "required": [
          "chapter",
          "dimensions"
        ],
        "title": "ChapterVerdict",
        "type": "object"
      },
...
      "DimensionOut": {
        "properties": {
          "dimension": {
            "title": "Dimension",
            "type": "string"
          },
          "discarded_instances": {
            "default": [],
            "description": "Instancias del Jurado sin puntuacion anclada en esta dimension (RF-129)",
            "items": {
              "type": "string"
            },
            "title": "Discarded Instances",
            "type": "array"
          },
          "dispersion": {
            "title": "Dispersion",
            "type": "integer"
          },
          "level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/ScoreOut"
            },
            "title": "Scores",
            "type": "array"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "dimension",
          "level",
          "dispersion",
          "valid",
          "scores"
        ],
        "title": "DimensionOut",
        "type": "object"
      },
...
      "FingerprintRow": {
        "description": "RD-21. La huella de un capitulo congelado.",
        "properties": {
          "adj_noun_ratio": {
            "title": "Adj Noun Ratio",
            "type": "number"
          },
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "deviation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviation"
          },
          "is_reference": {
            "default": false,
            "title": "Is Reference",
            "type": "boolean"
          },
          "lexical_richness": {
            "title": "Lexical Richness",
            "type": "number"
          },
          "mean_sentence_len": {
            "title": "Mean Sentence Len",
            "type": "number"
          },
          "top_ngrams": {
            "items": {
              "type": "string"
            },
            "title": "Top Ngrams",
            "type": "array"
          },
          "var_sentence_len": {
            "title": "Var Sentence Len",
            "type": "number"
          }
        },
        "required": [
          "chapter",
          "mean_sentence_len",
          "var_sentence_len",
          "adj_noun_ratio",
          "top_ngrams",
          "lexical_richness"
        ],
        "title": "FingerprintRow",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScoreOut": {
        "properties": {
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "scene_id": {
            "title": "Scene Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "scene_id",
          "instance",
          "seed",
          "score",
          "quote",
          "offset"
        ],
        "title": "ScoreOut",
        "type": "object"
      },
...
      "StyleReport": {
        "properties": {
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/FingerprintRow"
            },
            "title": "Chapters",
            "type": "array"
          },
          "novel_id": {
            "title": "Novel Id",
            "type": "string"
          }
        },
        "required": [
          "novel_id",
          "chapters"
        ],
        "title": "StyleReport",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "verification__routes__ErrorDetail": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Sistema autonomo de generacion de novelas largas. La API observa y arranca; no aprueba, no corrige y no desbloquea.",
    "title": "Story Maker",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/novels/{novel_id}/chapters/{number}/verdict": {
      "get": {
        "description": "RI-29. El veredicto del Jurado sobre un capitulo congelado.",
        "operationId": "get_verdict_novels__novel_id__chapters__number__verdict_get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "number",
            "required": true,
            "schema": {
              "title": "Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterVerdict"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/verification__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/verification__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Verdict",
        "tags": [
          "verification"
        ]
      }
    },
...
    "/novels/{novel_id}/style": {
      "get": {
        "description": "RI-30. Huella de referencia y la de cada capitulo, con su desviacion.",
        "operationId": "get_style_novels__novel_id__style_get",
        "parameters": [
          {
            "description": "Identificador de la novela",
            "in": "path",
            "name": "novel_id",
            "required": true,
            "schema": {
              "description": "Identificador de la novela",
              "pattern": "^[a-z0-9][a-z0-9-]{2,63}$",
              "title": "Novel Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StyleReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/verification__routes__ErrorDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/verification__routes__ErrorDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Style",
        "tags": [
          "verification"
        ]
      }
    },
...
  }
}